Upgrade to .Net Standard 2.1
Updates the target framework for test project to .NET 8.0.

Updates the target framework for the main project to netstandard2.1 and sets LangVersion to 10.

Updates package references for Polly and Microsoft.Extensions.Configuration to the latest versions and System.IdentityModel.Tokens.Jwt.

Adds IsExternalInit.cs to support record types in older target frameworks.